pyproject.toml
[project]
name = "mos_env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.14"
dependencies = [
    "torch==2.3.1+cu121",
    "torchvision==0.18.1+cu121",
    "torchaudio==2.3.1+cu121",
    "matplotlib==3.9.2",
    "inflect==7.3.1",
    "scipy==1.14.1",
    "soundfile==0.12.1",
    "tensorboard==2.17.1",
    "tqdm==4.66.5",
    "unidecode==1.3.8",
    "g2pm==0.1.2.5",
    "g2p-en==2.1.0",
    "pyyaml==6.0.2",
    "notebook==7.2.1",
    "einops==0.8.0",
    "sounddevice==0.5.0",
    "cython==3.0.11",
    "wandb==0.17.7",
    "pydub==0.25.1",
    "llvmlite==0.43.0",
    "librosa==0.10.2.post1",
    "onnx==1.16.2",
    "pandas==2.2.3",
    "seaborn==0.13.2",
    "jiwer==3.0.5",
    "pystoi==0.4.1",
    "fastdtw==0.3.4",
    "pyworld==0.3.4",
    "pysptk==1.0.1",
    "monotonic-align",
    "openai-whisper",
    "streamlit>=1.41.1",
]

[tool.uv]
find-links = [
    "https://download.pytorch.org/whl/cu121/torch",
    "https://download.pytorch.org/whl/cu121/torchvision",
    "https://download.pytorch.org/whl/cu121/torchaudio",
]

[tool.uv.sources]
monotonic-align = { git = "https://github.com/unrea1-sama/monotonic_align.git" }
openai-whisper = { git = "https://github.com/openai/whisper.git" }
